// src/element/element.interface.ts
/*
export interface InputAttributeInterface {
  autocomplete?: 'off' | 'on';
  disabled: boolean;
  maxlength?: number;
  placeholder?: string;
  required?: boolean;
  type: 'text';
}

export interface SelectAttributeInterface {
  autocomplete?: 'off' | 'on';
  disabled: boolean;
  maxlength?: number;
  placeholder?: string;
  required?: boolean;
  type: 'text';
}


export interface TypeOptionsInterface {
  key: string;
  model?: Object;
  attribute?: InputAttributeInterface | SelectAttributeInterface;
}
*/